- ExternalUserInformationProviderFactory is now scoped as singleton instance in Startup.cs - moved contracts and implementations to new \Providers\ExternalUserInformationProviders folder
